Module MenhirSdk.Cmly_api

module type INDEXED = sig ... end

The module type INDEXED describes a type t whose elements are in a bijection with an integer interval of the form [0..count).

module type GRAMMAR = sig ... end

The module type GRAMMAR describes the grammar and automaton.